# Optimize jpeg decoder based on traces.