--- a/src/ImageSharp/Formats/Jpeg/Components/Decoder/HuffmanScanBuffer.cs
+++ b/src/ImageSharp/Formats/Jpeg/Components/Decoder/HuffmanScanBuffer.cs
@@ -93,25 +93,24 @@ namespace SixLabors.ImageSharp.Formats.Jpeg.Components.Decoder
         public unsafe int DecodeHuffman(ref HuffmanTable h)
         {
             this.CheckBits();
-            int v = this.PeekBits(JpegConstants.Huffman.LookupBits);
-            int symbol = h.LookaheadValue[v];
-            int size = h.LookaheadSize[v];
+            int index = this.PeekBits(JpegConstants.Huffman.LookupBits);
+            int size = h.LookaheadSize[index];
 
-            if (size == JpegConstants.Huffman.SlowBits)
+            if (size < JpegConstants.Huffman.SlowBits)
             {
-                ulong x = this.data << (JpegConstants.Huffman.RegisterSize - this.remainingBits);
-                while (x > h.MaxCode[size])
-                {
-                    size++;
-                }
+                this.remainingBits -= size;
+                return h.LookaheadValue[index];
+            }
 
-                v = (int)(x >> (JpegConstants.Huffman.RegisterSize - size));
-                symbol = h.Values[(h.ValOffset[size] + v) & 0xFF];
+            ulong x = this.data << (JpegConstants.Huffman.RegisterSize - this.remainingBits);
+            while (x > h.MaxCode[size])
+            {
+                size++;
             }
 
             this.remainingBits -= size;
 
-            return symbol;
+            return h.Values[(h.ValOffset[size] + (int)(x >> (JpegConstants.Huffman.RegisterSize - size))) & 0xFF];
         }
 
         [MethodImpl(InliningOptions.ShortMethod)]